Rename `core` to `services`.
