+++ packages/blockchain-wallet-v4-frontend/src/modals/Btc/VerifyMessage/services.js
