+++ packages/blockchain-wallet-v4-frontend/src/modals/Btc/VerifyMessage/services.spec.js
@@ -1,15 +1,15 @@
-import * as core from './core'
+import * as services from './services'
 
 describe(`showResult`, () => {
   it(`should return false when all inputs are empty`, () => {
     expect(
-      core.showResult({ address: ``, message: ``, signature: `` })
+      services.showResult({ address: ``, message: ``, signature: `` })
     ).toEqual(false)
   })
 
   it(`should return false when some inputs are empty`, () => {
     expect(
-      core.showResult({
+      services.showResult({
         address: `address`,
         message: `message`,
         signature: ``
@@ -19,7 +19,7 @@ describe(`showResult`, () => {
 
   it(`return true when all inputs are provided`, () => {
     expect(
-      core.showResult({
+      services.showResult({
         address: `address`,
         message: `message`,
         signature: `signature`
@@ -31,7 +31,7 @@ describe(`showResult`, () => {
 describe(`verifySignature`, () => {
   it(`good signature`, () => {
     expect(
-      core.verifySignature({
+      services.verifySignature({
         address: `15WJg3bnKkuzeLxCKUGCKjRM42d3LgWg9u`,
         message: `Fun is nuf spelled backwards!`,
         signature: `HxB1FMYnYLda3SDhFVXpfyUfxfGOUOuCkrY0rwxIFaXzWgv/y7Wlijd9C/2t6ydoVOpi4v5XedgvDDOb8a6ZCfo=`
@@ -41,7 +41,7 @@ describe(`verifySignature`, () => {
 
   it(`bad signature`, () => {
     expect(
-      core.verifySignature({
+      services.verifySignature({
         address: `15WJg3bnKkuzeLxCKUGCKjRM42d3LgWg9u`,
         message: `Fun is nuf spelled backwards!`,
         signature: `bad signature`
